+++ src/io_mdds.h
@@ -75,9 +75,9 @@ class MEDDLY::mdd_reader {
         /// Create the forest based on the file
         mdd_reader(input &s, domain* D);
         void readRootEdge(dd_edge &E);
-        inline unsigned numRoots() const        { return roots.size(); }
-        inline forest* getForest() const        { return For; }
-        inline unsigned getFileNodes() const    { return file_nodes; }
+        inline unsigned numRoots() const            { return roots.size(); }
+        inline forest* getForest() const            { return For; }
+        inline size_t getFileNodes() const          { return file_nodes; }
         inline const dd_edge& getRoot(unsigned i) const
         {
             return roots[i];
@@ -89,7 +89,7 @@ class MEDDLY::mdd_reader {
         forest* For;
         std::vector <dd_edge> roots;
         unsigned rptr;
-        unsigned file_nodes;
+        size_t file_nodes;
 };
 
 #endif
